requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
langchain_community
langchain_openai
langchain_experimental
duckduckgo-search
SQLAlchemy
wikipedia
cassio
tiktoken
gradio
openai
PyPDF2
gradio
langgraph
faiss-cpu
langchain_groq